Set up queue and join game
I tried doing the queue and join process outside of the Rockpaperawesome React component, then passing the game in as a property, and storing it in the component's state.

However, I found out that Objects are not valid React children, which I think means you can't pass them in as props.