Start implementing invitation system
These additions start creating a system that allows users to create game invitations.

The final implementation will work as follows;
In invitation, a link with an embedded token is generated and given to the player (P1).
P1's id and the token are stored in the InvitationChannel
This player sends this link to a second player (P2), who opens the link.
Then P2's browser uses the token to connect to the InvitationChannel, with the token
specifying the topic.
If the server finds that token in memory, P1 and P2 are put into a game.