The best I could do
In order to get matrix animations working properly,
I had to add a keyframe between 45% and 90%. The rotation between these two wasn't going in the direction I wanted, so I found the keyframe percentage where one of the coordinates would change sign but the other wouldn't, so that only one sign change occurred between any two keyframes.
It sort of worked, but produced a slight jerky behavior around the new keyframe. I'd better just stick with the other transformation methods :/

I'm committing this even thought I'm not going to be using it.